analytics-data-platform/requirements.txt
# Blockchain SDKs
xrpl-py>=2.0.0
xahau-py>=0.1.0
web3>=6.0.0

# API Clients
pycoingecko>=3.0.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0

# Network & System Monitoring
speedtest-cli>=2.1.0
ping3>=4.0.0
websockets>=12.0

# Configuration Management
python-dotenv>=1.0.0

# Utilities